+++ concordium-node/build.rs
@@ -229,6 +229,15 @@ fn build_grpc2(proto_root_input: &str) -> std::io::Result<()> {
         )
         .method(
             tonic_build::manual::Method::builder()
+                .name("get_lock_info")
+                .route_name("GetLockInfo")
+                .input_type("crate::grpc2::types::LockInfoRequest")
+                .output_type("Vec<u8>")
+                .codec_path("crate::grpc2::RawCodec")
+                .build(),
+        )
+        .method(
+            tonic_build::manual::Method::builder()
                 .name("get_account_list")
                 .route_name("GetAccountList")
                 .input_type("crate::grpc2::types::BlockHashInput")
@@ -241,6 +250,16 @@ fn build_grpc2(proto_root_input: &str) -> std::io::Result<()> {
             tonic_build::manual::Method::builder()
                 .name("get_token_list")
                 .route_name("GetTokenList")
+                .input_type("crate::grpc2::types::BlockHashInput")
+                .output_type("Vec<u8>")
+                .codec_path("crate::grpc2::RawCodec")
+                .server_streaming()
+                .build(),
+        )
+        .method(
+            tonic_build::manual::Method::builder()
+                .name("get_lock_list")
+                .route_name("GetLockList")
                 .input_type("crate::grpc2::types::BlockHashInput")
                 .output_type("Vec<u8>")
                 .codec_path("crate::grpc2::RawCodec")