+++ plt/plt-scheduler-types/src/types.rs
@@ -2,6 +2,7 @@
 
 pub mod events;
 pub mod execution;
+pub mod locks;
 pub mod queries;
 pub mod reject_reasons;
 pub mod tokens;