+++ plt/plt-scheduler/tests/plt_token_list_updates.rs
@@ -164,6 +164,7 @@ fn test_add_allow_list_reject_non_governance() {
     let result = scheduler::execute_transaction(
         sender,
         sender_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -215,6 +216,7 @@ fn test_remove_allow_list_reject_non_governance() {
     let result = scheduler::execute_transaction(
         sender,
         sender_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -265,6 +267,7 @@ fn test_add_deny_list_reject_non_governance() {
     let result = scheduler::execute_transaction(
         sender,
         sender_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -315,6 +318,7 @@ fn test_remove_deny_list_reject_non_governance() {
     let result = scheduler::execute_transaction(
         sender,
         sender_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -471,6 +475,7 @@ fn test_add_to_not_enabled_allow_list() {
     let result = scheduler::execute_transaction(
         gov_account,
         stub.account_canonical_address(&gov_account),
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -514,6 +519,7 @@ fn test_remove_from_not_enabled_allow_list() {
     let result = scheduler::execute_transaction(
         gov_account,
         stub.account_canonical_address(&gov_account),
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -557,6 +563,7 @@ fn test_add_to_not_enabled_deny_list() {
     let result = scheduler::execute_transaction(
         gov_account,
         stub.account_canonical_address(&gov_account),
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -600,6 +607,7 @@ fn test_remove_from_not_enabled_deny_list() {
     let result = scheduler::execute_transaction(
         gov_account,
         stub.account_canonical_address(&gov_account),
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -644,6 +652,7 @@ fn test_reject_add_denylist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -663,6 +672,7 @@ fn test_reject_add_denylist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        2.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -710,6 +720,7 @@ fn test_reject_add_allowlist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -729,6 +740,7 @@ fn test_reject_add_allowlist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        2.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -784,6 +796,7 @@ fn test_reject_remove_denylist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -803,6 +816,7 @@ fn test_reject_remove_denylist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        2.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -858,6 +872,7 @@ fn test_reject_remove_allowlist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -877,6 +892,7 @@ fn test_reject_remove_allowlist_without_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        2.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -927,6 +943,7 @@ fn test_succeeds_add_deny_list_new_account_with_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -946,6 +963,7 @@ fn test_succeeds_add_deny_list_new_account_with_role() {
     let result = scheduler::execute_transaction(
         account2,
         account2_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -988,6 +1006,7 @@ fn test_succeeds_add_allow_list_new_account_with_role() {
     let result = scheduler::execute_transaction(
         gov_account,
         gov_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),
@@ -1007,6 +1026,7 @@ fn test_succeeds_add_allow_list_new_account_with_role() {
     let result = scheduler::execute_transaction(
         account2,
         account2_addr,
+        1.into(),
         stub.state_mut(),
         Payload::TokenUpdate { payload },
         Energy::from(u64::MAX),